Add flag forcing non-GUI JVM to exit after test
This is useful for cases when JMeter is run by an automated system
(such as CI builds) and tests start threads that prevent the JVM
from cleanly shutting down. The existing property to do a full JVM
shutdown (jmeterengine.stopfail.system.exit) only executes when the
test is explicitly asked to stop. We needed to unconditionally insure
the JVM is terminated when the test was finished, even if the tests
themselves did things which left the JVM in a state where JMeter could
not cleanly exit in non-GUI mode.

The new flag is "jmeterengine.force.system.exit" and defaults to
"false".