Bug 57394: When constructing an instance using ClassTools#construct(String, int), the int parameter was not used, but the default constructor was used instead.
Added a few test cases to test the methods of ClassTools and correct the unused parameter.

The only place in jmeter, where this method was called is in the JMeter class itself and
is used, when the property "mirror.server.port" is set to a value greater than zero. 
Without this patch that value would be ignored.

Bugzilla Id: 57394



git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1647607 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 8594db3e520320b54099987a10a2c22d04896984