Bug 52073 Embedded Resources concurrent download : Improve performances by avoiding shutdown of ThreadPoolExecutor at each sample.
A shared thread pool is used for all the parallel downloads.
A throttle mechanism is used per user to ensure the max concurrent
downloads setting is not exceeded.
A new parameter httpsampler.parallel_download_thread_keepalive_inseconds
is used to set the thread TTL when idle (default 60s).
the parameter httpsampler.await_termination_timeout is not used anymore
and has been removed.

TODO : ResourcesDownloader#shrink should be called at the end of a test
to accelerate the release of the thread pool idle threads (not
mandatory)

the httpclient is shared between the threads of a given virtual user.
the cache is shared between the threads of a given virtual user.