Bug59328_BetterLabelAndToolTipInCSVDataSetVariableNames
Hi,

in CSVDataSet if the variable name field is empty, then the first line
of the file is read and interpreted as the list of column names.

It's a great feature but we need to read the doc to have the
information.

I propose to put this information in the tooltip.

Second modification is to remove "Separate by commas" because this
information is already in the display name.

Antonio