Bug59349_IncludeControllerFilenameTrim
Hi,

Like I have proposed in "Bug 56877 - CSVDataSet does not trim spaces in
<filename>.csv", I split this issue into different PR/Issues to merge
fix step by step

With this fix I propose to trim spaces in input filename in
IncludeController.

Without this fix, if the user put a space at the beginning of the input
filename, we have an exception 

It's added to "Bug 59329 - Trim spaces in input filename in CSVDataSet".

Very soon next PR to complete Bug 56877


Antonio