Supply Subject Alternative Name on HTTPS certs.
Recently browsers have started to not trust a certificate if it doesn’t have a Subject Alternative Name (SAN) that matches the domain name. This causes the HTTPS proxy to generate warnings in the browser even if the CA certificate is trusted.

This sets the SAN on all the certificates that are generated so that the browsers trust them.