Use the wrapped exception for error message on file load for StreamException
When a user opens a test plan that is no xml at all, a StreamException
gets thrown. This exception can have an empty message. That will lead
to an error dialog that has no text at all. In that case we try to work
around this by using the wrapped exception as the source for the error
message.

Bugzilla Id: 62683


git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1840238 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: a8e5d3b1f7a46e14820e01374010c00081fc37ae