Allow full JSR-223 for init scripts in JMeter startup
Use jsr223.init.file to allow calling a JSR-223 script on JMeter startup
instead of the envisioned groovy.init.file property.

The engine to use for the init script will be determined by the extension
of the filename. If the filename has no extension or if no engine
can be found for that extension, Groovy will be tried.

Bugzilla Id: 62700


git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1840847 13f79535-47bb-0310-9956-ffa450edef68