Use simpler code for "static" String concatenation
Java is smart enough to compile simple string concatenation to
the same code as `new StringBuilder("something").append("other thing").toString()`.
The latter is really hard to read, so use the simpler one.

While at it, use string formats for the logger (as it is some sort of string
concatenation, too). And since we use the string formatter, we don't have
to guard the log level anymore for the simple case.