Add missing License header
Followup to r1844031

Render uninitialized min and max values in Summary Report as `#N/A`

In the Summary Report min and max values are initialized with
`Long#MAX_VALUE` and `Long#MIN_VALUE` to work easily with them.

Those initial values are looking weird in the rendered view.
Hide them by replacing the extrema by the text `#N/A`.

Closes #407 on github
Bugzilla Id: 62822



git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1844032 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 6f6fa40c3c684106cf1b54d4ae207e8b829a1314