Bug 59896 - Report / Dashboard : Add a menu entry to generate a report on demand from a CSV file
There is a bug in the jdk on X11, that keeps the dialog on screen when setVisible(false) is called too early after
the dialog appeared on screen.

See https://bugs.openjdk.java.net/browse/JDK-5109571 for more information

It would probably be better to not show the dialog in the first place, when we are testing the component.

Relates to #453
Bugzilla Id: 59896



git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1855100 13f79535-47bb-0310-9956-ffa450edef68