Correct synchronized usage of field
Most of the time the file keyStore was already used inside a synchronized block and volatile
is not a good idea to be used for non primitive types, so use synchronized for the rest of the
simple methods that use keyStore as field.

Followup to r1856123. Now remove the volatile keyword, really.



git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1856124 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 8e2643b4412779b1118579635cf43316e7acb9be