Use a different way to calculate a working IP address for Bug52310
Many modern linux systems are configured to localhost for the first network adapter.
This will break the test plan for Bug52310. Now we are trying to get the first
IPv4 address that is not a loopback address.

Use groovy instead of beanshell to do the calculation.

Remove HC3.1 reference, as JMeter doesn't support this anymore. Use Java client instead,
even if it was not the one that had the bug in the first place.


git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1859990 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 019a9d582549fa1509f51b96717aca4ebaed343d