Revert behaviour of TimerService#adjustDelay to that of 5.1.1
Changing the adjustDelay method to return -1 when the delay would be longer
than the remaining scheduled duration, lead to bug 63711.

This change reverts the default adjustDelay to the old behaviour and adds a
flag to enable the "new" feature of returning -1 when the sleep time would
be too long.

Bugzilla Id: 63711