Use a sane initial dir for the file chooser
While having a look at 63945, I found a few places, where we look for the existence
of the initial directory on which the file chooser should start.

I extended and simplified those occurrences and think I found a bug in the logic that
chooses the initial location in FileDialoger#promptToOpenFile from line 177.

First we look, if existingFileName is pointing to something valid and use that as the
initial location or the global initial location, if none location has been used before.

After that a bit further down, we test if a location has been used before and may
initialize such a location.

Than -- and this is the potential bug -- we always use that location for the initial
location of the file chooser.

Another dodgy logic is that we test for the default initial location in two different ways.
First -- when no existingFileName is given -- we test for an empty default location and refuse
to use it, if it is empty. The second time, we accept the default location, even if it is
empty AND use it as the default location for the next runs.

Bugzilla Id: 63945