tz-magic when dst-transition is near
Our tests (for timeshift function) started to fail shortly before a DST transition in
New Yorker timezone, as the parsing of a date into LocalDateTime seems to follow
different rules than adding time to a LocalDateTime instance.

Therefore we now try to calculate the DST offset for the calculated time and the current
time. That offset will be added to the result of the timeshift function.