Recurse into sampleResults for getSampleList
With this change we collect the sampleList of a SampleReult by recursing up to
three levels into sub results of SampleResults instead of only one level deep.

This helps in case of parallel downloads of embedded elements of a web page, which
get stored as sub-sub-sampleresults. Before this change a regex Post Processor would
not see those sub-sub-sampleresults.

Bugzilla Id: 64196