Initialize GUI_CLASS in TestBeanGUI so the icons are rendered properly
Since 0bdbe5bd17f6e2385a5fe58216e259dd85a04054 JMeter skips
TestBeanGUI#getMenuCategories() for most of the beans,
and it turns out that method was initializing GUI_CLASS which is used
for icon rendering.

The solution for now is to explicitly initialize GUI_CLASS when TestBeanGUI's
UI is first initialized.