Update JMeterTreeModel in a threadsafe and timely manner
When recording multiple samples concurrently, we could get into a race
condition and place samples into wrong controllers.

Now we try to get the current information about prefix and sampler as soon as possible
and delay the addition of the sampler to the recording controller by means of a queue.