Freestyle format for names in (Default)SamplerCreater
Names of newly created Samplers by the DefaultSamplerCreator can now be
styled with a freestyle format, that understands the controls for
MessageFormat. Apart from the placeholders {0}, {1} and {2}, that are
hard to guess, for which part they are used, more memorizable placeholders
 #{name}, #{path} and #{counter} can be used.

The counter for the generated samplers can be set to a user specified number.

Based on a patch by Vincent Daburon

Bugzilla Id: 64696
Closes #571
Closes #595