Implement a new setting to allow the exclusion of embedded URLs
At the moment we allow only one regular expression to control which URLs will
be downloaded by the HTTP sampler, when embedded download is enabled.

This can be used to exclude specific URLs, but it is not possible to exclude and allow
URLs at the same time.

This patch gives the user a second field which can be used to exclude URLs.

Bugzilla Id: 63527