Don't assume we have correct JSON at edit time
When JMeter variables are used inside variable or query blocks for the
GraphQL Sampler, we cannot assume, that we still have valid JSON as
long those variables are not rendered. The rendering of the variables
is done at a later time in the life cycle. Therefore, we have to do
our best to render JSON without actually checking the validity.

Bugzilla Id: 65108