JSONPathAssertion attributes are out of order
Compare JSON objects and not their string representations.

When using our stringifier, the order of the entries in maps
is not guaranteed and can lead to wrong results.

In the old days we made no difference between a string or int
when asserting a result. Jackson JSON Parser differentiates
between 'foo' and '"foo"' (former is invalid) and '1' and '"1"'
(former is an int, latter a string).

To enable both (complex, simple and edge cases), we now have to do
more work.

Bugzilla Id: 65299