Restart does not work when parameters contain spaces
Try to re-combine the parameters to regard spaces in file names

Java command line has no knowledge of the 'real' parameters and
we have to do a bit of guessing to re-assemble the parameters with
spaces and the drop the spaces, that should split the parameters.

So we guess, that each parameter starts with a dash (-) and
everything else are values, that should be stitched together.

Bugzilla Id: 65522