Use default values for null values when extracting with JSONPostProcessor
When we extract a null value for a given expression, we now replace that
null with the given default value for the expression.

That leads to slight change of the meaning of a missing value.

Consider the JSON structure: {"context": null} and the JSON Path: "$.context"

The new implementation will replace the `null` with the default value, while
the old would have stringified it to the empty string.

Cleanup of parameter handling of default values. We should use the default
value for the current handled expression and not pass all default values
for all expression. That leads to confusion, only. (Well, I was confused)

Bugzilla Id: 65681