Fail JSONPath Assertion on indefinite paths and no assertion value
The docs for JSONPath Assertion state, that it will fail the assertion,
when no element is found with the given JSON path. This was currently
not followed, when an indefinite path was used. In such a case, the
JSONPath library would return an empty list, which the assertion logic
would log as OK.

With this change we let the assertion fail, when
 * an indefinite path was given
 * and an empty list is extracted
 * and no assertion value is given

Bugzilla Id: 65794