Enable passing of system properties to tests via special namespace
If a system property prefixed with jmeter.properties. is found, Gradle will
now pass it as a system property without that prefix to the unit tests.
That allows to set system properties (which might be used as jmeter properties)
to the unit tests, without setting them in the JVM, that is used by Gradle.