Get rid of error prone warnings because of missing explicit default character encoding
While we are in the sources, use try-with, isEmpty and other minor clean-ups.
Don't encode ASCII printable characters as unicode sequences, as they are
decoded by Java anyway and are obfuscating the code unnecessarily.
Remove unused method variables and parameters, when possible.
Use ArrayList instead of LinkedList, when possible (more will follow).