Silence warnings in test code and use correct parameter when calling chained constructor
When chaining the constructor to add a default UA, we lost one parameter (replaced it
with another). That seems wrong, so I changed it back to the (hopefully) correct one.

While we are here, suppress a few other warnings from error prone, that can't be
fixed in a quick way. Feel free, to fix them.