Use reference equality in AbstractTestElement.equals, and add TestElement.contentEquals
Previously, AbstractTestElement.hashCode used System.identityHashcode,
so it was not consistent with equals, and TestElements were often used in
HashMap keys.

So the fix is to use reference equality in equals, so it becomes consistent
with hashCode.

AbstractTestElement.equals is marked deprecated because so the clients could
adjust their code if they really need compare the contents of the TestElements
rather than references.