Warn about time format change only once
There could be other reasons, why 'u' is contained in the
format string. So use the format with a known date and check
the result, if 'u' might have been used.

The warning will be issued only once.

This implementation has a few loopholes/problems:
 * 'u' could still be used for something different and the year
   could be coming from a 'YYYY' format string
 * we add a volatile variable, which adds overhead

I start to believe that it is probably better to revert to
the old logic and live with the IDE warnings.

Part of #5694