test: Add OpenJDK JMH for creating microbenchmarks in JMeter code
This enables creating benchmark for tracking performance of low-level code.

The way to add a benchmark is to add src/jmh folder (side by side with src/main).
Then benchmark jar can be generated with ./gradlew jmhJar, or the benchmark
can be executed from IDE if there's a corresponding main method.

See ResultCollectorBenchmark