fix: restore public static final String references to literals for backward compatibility
Previously, TestElement.NAME and similar constants were coded as literals,
and it was possible to use those references in annotation values.

If the replace literals with method calls, then it would be impossible to use
the fields in annotation values.

So the static fields are reverted to literals.

This partially reverts commits like https://github.com/apache/jmeter/pull/678,