feat: Add checkbox to enable/disable params in HTTP Sampler
Added a new "isEnabledFromGui" property in HTTPArgument used
to determine whether to include the argument in the request
or not.
This property is set by the user for each HTTP parameter by
clicking on the "Enable" column put on the very left of each
parameter row.

Fixes https://github.com/apache/jmeter/issues/5466