Fix KeyManage log when using CLI mode
Without this patch, whenever a KeyStore is used (for example for client
certificates) in CLI mode, there is a message that no password was
provided.
This was caused by the else of the condition being on the wrong if
statement, and can cause people to spend a lot of time trying to debug
why the key store won't work.