Add explanation to UnrecoverableKeyException
Unrecoverable keys after opening the keystore could be a consequence of
the key having a different password than the keystore itself.
This could cause some people to spend a lot of time figuring out what's
going on, so let's just give them a quick heads up of a potential
(common?) reason for this error.

This could be fixed by running "keytool -keypasswd".