chore: use 127.0.0.1 or ::1 for client-server tests
InetAddress.getLocalHost().canonicalHostName might resolve to a public NAT
DNS name, and there will be no way to bind a server port for it.

Here's an example how the bind fails in Apache Jenkins EC2 instance:

[server] Server failed to start: java.rmi.server.ExportException: Listen failed on port: 0; nested exception is:
[server] 	java.io.IOException: Could not bind to bb-worker1-ec2-va.apache.org/3.234.254.195 using port 0

The workaround is to avoid InetAddress.getLocalHost() altogether,
so we use 127.0.0.1 when IPv4 is available or ::1 otherwise.